+++ lib/brando_admin/channels/lobby_channel.ex
@@ -15,36 +15,8 @@ defmodule Brando.LobbyChannel do
   Join lobby channel
   """
   def join("lobby", _params, socket) do
-    list_opts = %{select: [:id, :name], cache: {:ttl, :infinite}, preload: [{:avatar, :join}]}
-    {:ok, users} = Brando.Users.list_users(list_opts)
-
-    users =
-      Enum.map(
-        users,
-        fn user ->
-          avatar =
-            if user.avatar && Brando.Images.Utils.image_type(user.avatar.path) == :svg do
-              Brando.Utils.img_url(user.avatar, :original,
-                prefix: "/media",
-                default: "/images/admin/avatar.svg"
-              )
-            else
-              Brando.Utils.img_url(user.avatar, :smallest,
-                prefix: "/media",
-                default: "/images/admin/avatar.svg"
-              )
-            end
-
-          %{
-            name: user.name,
-            id: user.id,
-            avatar: avatar
-          }
-        end
-      )
-
     send(self(), :after_join)
-    {:ok, %{users: users}, socket}
+    {:ok, %{}, socket}
   end
 
   def handle_in("user:state", %{"active" => active}, socket) do
@@ -73,7 +45,6 @@ defmodule Brando.LobbyChannel do
         url: nil
       })
 
-    push(socket, "presence_state", Brando.presence().list(socket))
     {:noreply, socket}
   end
 end