+++ lib/brando_admin/components/form.ex
@@ -36,6 +36,8 @@ defmodule BrandoAdmin.Components.Form do
      |> assign(:file_changeset, nil)
      |> assign(:image_changeset, nil)
      |> assign(:initial_update, true)
+     |> assign(:dirty_fields, [])
+     |> assign(:presences, %{})
      |> assign(:has_meta?, false)
      |> assign(:status_revisions, :closed)
      |> assign(:processing, false)
@@ -542,6 +544,8 @@ defmodule BrandoAdmin.Components.Form do
               </button>
             </div>
 
+            <.form_presences presences={@presences} />
+
             <div class="form-tab-builtins">
               <button
                 :if={@has_meta?}
@@ -692,6 +696,22 @@ defmodule BrandoAdmin.Components.Form do
           live_preview_target={@live_preview_target}
           change_preview_target={JS.push("change_preview_target", target: @myself)}
         />
+      </div>
+    </div>
+    """
+  end
+
+  attr :presences, :list
+
+  def form_presences(assigns) do
+    ~H"""
+    <div class="page-presences">
+      <div
+        :for={{_, user} <- @presences}
+        class="user-presence visible">
+        <div class="avatar">
+          <Content.image image={user.avatar} size={:thumb} />
+        </div>
       </div>
     </div>
     """
@@ -1328,7 +1348,8 @@ defmodule BrandoAdmin.Components.Form do
             current_user: current_user,
             singular: singular,
             live_preview_active?: live_preview_active?,
-            live_preview_cache_key: live_preview_cache_key
+            live_preview_cache_key: live_preview_cache_key,
+            dirty_fields: dirty_fields
           }
         } = socket
       ) do
@@ -1336,6 +1357,21 @@ defmodule BrandoAdmin.Components.Form do
     entry_or_default = entry || struct(schema)
 
     changeset = validate(schema, entry_or_default, entry_params, current_user)
+
+    changed_fields = Map.keys(changeset.changes)
+
+    socket =
+      if changed_fields != dirty_fields do
+        Phoenix.PubSub.broadcast(
+          Brando.pubsub(),
+          "brando:dirty_fields:#{entry.id}",
+          {:dirty_fields, changed_fields, current_user.id}
+        )
+
+        assign(socket, :dirty_fields, changed_fields)
+      else
+        socket
+      end
 
     if live_preview_active? do
       Brando.LivePreview.update(schema, changeset, live_preview_cache_key)