+++ lib/brando_admin/live_view/form.ex
@@ -23,16 +23,6 @@ defmodule BrandoAdmin.LiveView.Form do
       def on_mount(:hooks, params, assigns, socket) do
         BrandoAdmin.LiveView.Form.hooks(params, assigns, socket, unquote(schema))
       end
-
-      # we need the uri on first load, so inject for now
-      def handle_params(params, url, socket) do
-        uri = URI.parse(url)
-
-        {:noreply,
-         socket
-         |> assign(:params, params)
-         |> assign(:uri, uri)}
-      end
     end
   end
 
@@ -48,6 +38,8 @@ defmodule BrandoAdmin.LiveView.Form do
         |> assign_entry_id(entry_id)
         |> set_admin_locale()
         |> attach_hooks(schema)
+
+      Phoenix.PubSub.subscribe(Brando.pubsub(), "brando:dirty_fields:#{entry_id}")
 
       {:cont, socket}
     else
@@ -79,18 +71,8 @@ defmodule BrandoAdmin.LiveView.Form do
 
   defp attach_hooks(socket, _schema) do
     socket
-    |> attach_hook(:b_form_params, :handle_params, &handle_params/3)
     |> attach_hook(:b_form_events, :handle_event, &handle_event/3)
     |> attach_hook(:b_form_infos, :handle_info, &handle_info/2)
-  end
-
-  defp handle_params(params, url, socket) do
-    uri = URI.parse(url)
-
-    {:halt,
-     socket
-     |> assign(:params, params)
-     |> assign(:uri, uri)}
   end
 
   defp handle_event(
@@ -116,6 +98,30 @@ defmodule BrandoAdmin.LiveView.Form do
   end
 
   defp handle_event(_, _, socket), do: {:cont, socket}
+
+  defp handle_info({:dirty_fields, fields, user_id}, socket) do
+    socket =
+      if user_id == socket.assigns.current_user.id do
+        Brando.presence().update_dirty_fields(socket.assigns.uri.path, user_id, fields)
+        socket
+      else
+        # TODO: there are updated dirty fields from other users.
+        require Logger
+
+        Logger.error("""
+
+        ==> dirty_fields
+
+        #{inspect(fields, pretty: true)}
+        #{inspect(user_id, pretty: true)}
+
+        """)
+
+        socket
+      end
+
+    {:halt, socket}
+  end
 
   defp handle_info({image, [:image, :updated], path}, socket) do
     case String.split(image.config_target, ":") do